# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventApp API",
        "description": "API REST pour la gestion d'événements collaboratifs - Plateforme de création, gestion et inscription à des événements",
        "version": "1.0.0",
        "contact": {
            "name": "EventApp Support",
            "email": "support@eventapp.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Inscription d'un nouvel utilisateur",
                "operationId": "post_api_register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firstName",
                                    "lastName"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 8,
                                        "example": "motdepasse123"
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Dupont"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur cr\u00e9\u00e9 avec succ\u00e8s",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                },
                                                "firstName": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "lastName": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Donn\u00e9es invalides"
                    },
                    "409": {
                        "description": "Email d\u00e9j\u00e0 utilis\u00e9"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Connexion utilisateur",
                "operationId": "post_api_login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion r\u00e9ussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                },
                                                "firstName": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "lastName": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants incorrects"
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "R\u00e9cup\u00e9rer le profil utilisateur",
                "operationId": "get_api_me",
                "responses": {
                    "200": {
                        "description": "Profil utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "user@example.com"
                                                },
                                                "firstName": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "lastName": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifi\u00e9"
                    }
                }
            }
        },
...
    },
...
}